+++ app/Http/Controllers/SupplierController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Supplier;
+use Illuminate\Support\Facades\Gate;
 
 class SupplierController extends Controller
 {
@@ -14,11 +15,15 @@ class SupplierController extends Controller
      */
     public function index()
     {
+        $this->authorize('supplier_index');
         return view('supplier.index');
     }
 
     public function data()
     {
+        if (Gate::denies('supplier_index')) {
+            return jsonResponse("Anda tidak dapat Mengakses Halaman atau Tindakan ini", 403);
+        }
         $supplier = Supplier::orderBy('id_supplier', 'desc')->get();
 
         return datatables()
@@ -28,7 +33,7 @@ class SupplierController extends Controller
                 return '
                 <div class="btn-group">
                     <button onclick="editForm(`'. route('supplier.update', $supplier->id_supplier) .'`)" class="btn btn-xs btn-info btn-flat"><i class="fa fa-pencil"></i></button>
-                    <button onclick="deleteData(`'. route('supplier.destroy', $supplier->id_supplier) .'`)" class="btn btn-xs btn-danger btn-flat"><i class="fa fa-trash"></i></button> 
+                    <button onclick="deleteData(`'. route('supplier.destroy', $supplier->id_supplier) .'`)" class="btn btn-xs btn-danger btn-flat"><i class="fa fa-trash"></i></button>
                 </div>
                 ';
             })
@@ -54,6 +59,9 @@ class SupplierController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('supplier_create')) {
+            return jsonResponse("Anda tidak dapat Mengakses Halaman atau Tindakan ini", 403);
+        }
         $supplier = Supplier::latest()->first();
         $supplier = Supplier::create($request->all());
         return response()->json('Data Berhasil Disimpan', 200);
@@ -67,6 +75,9 @@ class SupplierController extends Controller
      */
     public function show($id)
     {
+        if (Gate::denies('supplier_edit')) {
+            return jsonResponse("Anda tidak dapat Mengakses Halaman atau Tindakan ini", 403);
+        }
         $supplier = Supplier::find($id);
 
         return response()->json($supplier);
@@ -92,6 +103,9 @@ class SupplierController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (Gate::denies('supplier_edit')) {
+            return jsonResponse("Anda tidak dapat Mengakses Halaman atau Tindakan ini", 403);
+        }
         $supplier = Supplier::find($id);
         $supplier->nama_supplier = $request->nama_supplier;
         $supplier->alamat = $request->alamat;
@@ -110,6 +124,9 @@ class SupplierController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('supplier_delete')) {
+            return jsonResponse("Anda tidak dapat Mengakses Halaman atau Tindakan ini", 403);
+        }
         $supplier = Supplier::find($id);
         $supplier->delete();
